Smart_home_System/requirements.txt
paho-mqtt==1.6.1
redis==4.5.1
psycopg2-binary==2.9.6
